+++ src/components/ImageList/index.js
@@ -5,27 +5,18 @@ import {
   Image,
   TouchableOpacity,
   Dimensions,
-  Modal,
-  ActivityIndicator,
   Text,
 } from 'react-native';
-// import ImageViewer from 'react-native-image-zoom-viewer';
 import PropTypes from 'prop-types';
 
 import Img from '../Img';
 import { BackIcon } from '../Icons';
-import { trueTypeOf } from '../../utils';
 import styles from './styles';
 
 // added save area view dependenciy
 import { SafeAreaView } from 'react-native-safe-area-context';
 
 class ImageList extends Component {
-  state = {
-    imgToShow: 0,
-    showImg: false,
-  };
-
   componentDidMount() {
     const { index } = this.props;
     setTimeout(() => {
@@ -35,22 +26,18 @@ class ImageList extends Component {
     }, 0);
   }
 
-  showImg = () => {
-    const { showImg } = this.state;
-    this.setState({ showImg: !showImg });
-  };
-
-  showImgInModal = (indexOfImg) => {
-    // render the modal only when a children component was set
-    if (!this.props.children) return null;
-    this.setState({ imgToShow: indexOfImg, showImg: true });
+  handleItemPress = (indexOfImg) => {
+    // added custum even to inform the parent about the image press event
+    this.props.onImageListItemTap
+      ? this.props.onImageListItemTap(indexOfImg)
+      : undefined;
   };
 
   renderBackBtn = () => {
     const { setModalVisible, title } = this.props;
 
     return (
-      <SafeAreaView style={styles.backBtnView}>
+      <View style={styles.backBtnView}>
         <TouchableOpacity
           onPress={() => setModalVisible(false)}
           style={styles.backBtnTouchableOpacity}
@@ -58,11 +45,11 @@ class ImageList extends Component {
           <Image
             source={{ uri: BackIcon }}
             resizeMode="contain"
-            style={{ height: 30, marginLeft: -10 }}
+            style={styles.backBtnIcon}
           />
         </TouchableOpacity>
-        {title && <Text style={styles.title}>{title}</Text>}
-      </SafeAreaView>
+        {title && <Text style={styles.title}> {title}</Text>}
+      </View>
     );
   };
 
@@ -72,7 +59,7 @@ class ImageList extends Component {
       <TouchableOpacity
         key={`List-Image-${i + 1}`}
         style={[styles.clickableImg, { backgroundColor }]}
-        onPress={() => this.showImgInModal(i)}
+        onPress={() => this.handleItemPress(i)}
         activeOpacity={0.8}
       >
         <Img image={img} />
@@ -82,40 +69,23 @@ class ImageList extends Component {
   };
 
   render() {
-    const { showImg, imgToShow } = this.state;
-    const { images } = this.props;
-
-    const imageUrls = images.map((image) => {
-      const hasMoreData = trueTypeOf(image) === 'object';
-      return { url: hasMoreData ? image.url : image };
-    });
-
     return (
-      <View style={styles.container}>
+      <SafeAreaView style={styles.container}>
         {this.renderBackBtn()}
         <ScrollView ref={(view) => (this.scrollView = view)}>
           {this.renderImages()}
         </ScrollView>
-
-        <View style={styles.modalView}>
-          <Modal
-            visible={showImg}
-            transparent={false}
-            onRequestClose={() => this.setState({ showImg: false })}
-          >
-            {this.props.children ?? null}
-          </Modal>
-        </View>
-      </View>
+      </SafeAreaView>
     );
   }
 }
 
 ImageList.defaultProps = {
-  backgroundColor: 'lightgreen',
+  backgroundColor: 'transparent',
   index: 0,
   setModalVisible: () => {},
   title: undefined,
+  onImageListItemTap: () => {}, // share image list item click to parent
 };
 
 ImageList.propTypes = {
@@ -135,6 +105,7 @@ ImageList.propTypes = {
   index: PropTypes.number,
   setModalVisible: PropTypes.func,
   title: PropTypes.string,
+  onImageListItemTap: PropTypes.func,
 };
 
 export default ImageList;