+++ ChangeLog.md
@@ -8,11 +8,19 @@ All notable changes to this project will be documented in this file.
 
 - Changelog file.
 
+### Removed
+
+- Remove unecessery dependencies, and update react-native dependencies
+- Enhance the Item preview screen header, buttons and appearance
+- Add Custom ImageList onPress event to display a custom component or make a custom action on ImageList Item press
+- Remove the dependency to react-native-image-viewer package
+
 ### Fixed
 
 - Fix typos in README and usage.png file
 - Improved code readability
 - Decreased sample images' size
+- Fix close ImageList modal from the parent of the preview grid
 
 ## [1.0.0] - 2019-06-10